Fix behavior when Flush is called several times concurrently.
I noticed that when Flush was called from multiple goroutines
simultaneously, we'd occasionally see repeated log lines as the
LogDNA client would send the same payload twice. This is unlikely
when using the DefaultFlushLimit, but we'd encounter it when we
turned the FlushLimit down to 1. I fixed this by adding a mutex
to the Client such that only one Flush can occur at once for each
Client.

Additionally, I moved the check to see if the FlushLimit had been
reached to the end of Log(), so that if FlushLimit is 1, each
message is sent immediately as it is Logged.